Options for orphatools

Interface to list and set orphatools-related options.

Options are useful to roll back to older versions or to switch between languages for display.

"orphatools_dict" refers to the dictionary used to interprete Orphanet concepts ids.

"orphatools_nomenclature" refers to a nomenclature pack version (extraction date and used language).

"orphatools_gene_file" refers to the version of the file used to associated genes to ORPHAcodes.

All options can be set manually via the built-in options() function.

Usage

orphatools_options()

See also

Examples

if (FALSE) orphatools_options() # \dontrun{}